upd: 'auth completed' page added
In this page I set user data in localstorage with zustand and say user message like 'you may close this page' (because user have already opened one)
